+++ libs/common/src/lib/notifications/notification.model.ts
@@ -1,4 +1,6 @@
-export type NotificationType = 'resource-processing';
+import { NotificationOperation } from '@nuclia/core';
+
+export type NotificationType = 'resource';
 
 export interface NotificationData {
   title: string;
@@ -8,6 +10,7 @@ export interface NotificationData {
 
 export interface NotificationUI {
   type: NotificationType;
+  operation: NotificationOperation;
   timestamp: string;
   unread: boolean;
   failure: boolean;