+++ libs/common/src/lib/notifications/notification.service.ts
@@ -1,14 +1,17 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject, combineLatest, map, Observable, switchMap, take, tap } from 'rxjs';
+import { BehaviorSubject, combineLatest, filter, map, Observable, switchMap, tap } from 'rxjs';
 import { NotificationData, NotificationUI } from './notification.model';
 import { SDKService } from '@flaps/core';
 import { NavigationService } from '../services';
 import { differenceInSeconds } from 'date-fns';
+import { WritableKnowledgeBox } from '@nuclia/core';
+import { takeUntil } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
 })
 export class NotificationService {
+  private _currentKb?: WritableKnowledgeBox;
   private _notifications = new BehaviorSubject<NotificationUI[]>([]);
   notifications: Observable<NotificationUI[]> = this._notifications.asObservable();
 
@@ -19,11 +22,19 @@ export class NotificationService {
   constructor(
     private sdk: SDKService,
     private navigationService: NavigationService,
-  ) {
+  ) {}
+
+  startListening() {
     combineLatest([this.sdk.currentAccount, this.sdk.currentKb])
       .pipe(
+        tap(([, kb]) => {
+          if (this._currentKb) {
+            this.stopListening();
+          }
+          this._currentKb = kb;
+        }),
         switchMap(([account, kb]) =>
-          kb.listenToProcessingNotifications().pipe(
+          kb.listenToResourceOperationNotifications().pipe(
             tap((notifications) => {
               let existingNotifications = this._notifications.value;
               // most recent notifications are first in the list
@@ -37,10 +48,11 @@ export class NotificationService {
                     ? this.navigationService.getResourcePreviewUrl(account.slug, kb.slug, newNotif.resourceId)
                     : undefined,
                 };
-                // we group notifications happening within 30 seconds only for unread notification
+                // we group notifications happening within 30 seconds only for unread notification of the same operation type
                 if (
                   lastNotification &&
                   lastNotification.unread &&
+                  lastNotification.operation === newNotif.operation &&
                   differenceInSeconds(new Date(newNotif.timestamp), new Date(lastNotification.timestamp)) <= 30 &&
                   newNotif.success === !lastNotification.failure
                 ) {
@@ -48,7 +60,8 @@ export class NotificationService {
                   existingNotifications = [{ ...lastNotification }].concat(existingNotifications.slice(1));
                 } else {
                   newNotifications.push({
-                    type: 'resource-processing',
+                    type: 'resource',
+                    operation: newNotif.operation,
                     timestamp: newNotif.timestamp,
                     failure: !newNotif.success,
                     unread: true,
@@ -60,6 +73,7 @@ export class NotificationService {
             }),
           ),
         ),
+        takeUntil(this.sdk.nuclia.auth.isAuthenticated().pipe(filter((isAuthenticated) => !isAuthenticated))),
       )
       .subscribe();
   }
@@ -73,6 +87,10 @@ export class NotificationService {
   }
 
   stopListening() {
-    this.sdk.currentKb.pipe(take(1)).subscribe((kb) => kb.stopListeningToNotifications());
+    if (this._currentKb) {
+      this._currentKb.stopListeningToNotifications();
+      this._currentKb = undefined;
+      this._notifications.next([]);
+    }
   }
 }