+++ libs/sdk-core/src/lib/rest/rest.ts
@@ -1,6 +1,9 @@
 import { from, map, Observable, of, Subscriber, switchMap } from 'rxjs';
 import type { INuclia, IRest } from '../models';
 
+export const ABORT_STREAMING_REASON = 'Stop listening to streaming';
+const NS_BINDING_ABORTED_ERROR = 'TypeError: NetworkError when attempting to fetch resource.';
+
 /**
  * Handles the elementary REST requests to the Nuclia backend, setting the appropriate HTTP headers.
  *
@@ -331,14 +334,25 @@ export class Rest implements IRest {
         }
       },
       (reason) => {
-        // Error on fetch can be caused by the backend not closing gracefully the stream on time (causing errors like NS_ERROR_NET_PARTIAL_TRANSFER, or CORS error)
-        // If there was no error before, or last error was more than 10s ago, we reconnect
-        if (!this.streamErrorAt || Date.now() - this.streamErrorAt > 10000) {
-          this.streamErrorAt = Date.now();
-          this.fetchStream(path, observer, controller);
-        } else {
-          observer.error(reason);
+        // when aborting the fetch using the AbortController, we provide the ABORT_STREAMING_REASON
+        // allowing us to know we should not raise an error in the observer
+        if (reason === ABORT_STREAMING_REASON) {
           observer.complete();
+        } else {
+          // Error on fetch can be caused by the backend not closing gracefully the stream on time (causing errors like NS_ERROR_NET_PARTIAL_TRANSFER, or CORS error)
+          // If there was no error before, or last error was more than 10s ago, we reconnect
+          // except if the error reason is from NS_BINDING_ABORTED, which happens when reloading the page on firefox
+          if (
+            reason.toString() !== NS_BINDING_ABORTED_ERROR &&
+            (!this.streamErrorAt || Date.now() - this.streamErrorAt > 10000)
+          ) {
+            console.warn(`Message stream lost: "${reason}". Reconnecting at ${new Date()}`);
+            this.streamErrorAt = Date.now();
+            this.fetchStream(path, observer, controller);
+          } else {
+            observer.error(`Message stream lost: ${reason}`);
+            observer.complete();
+          }
         }
       },
     );